Module dock_core_rpc::did

Modules

did_method_key

offchain

onchain

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

signature

Structs

AddControllers

AddKeys

AddServiceEndpoint

AggregatedDidDetailsRequestParams

Information requested for DID. The default option includes only basic DID details.

AggregatedDidDetailsResponse

Aggregated details for the given DID.

Controller

DID’s controller.

Did

The type of the Dock DID.

DidKey

Valid did key with correct verification relationships.

DidKeyWithId

DidKey with its identifier.

DidMethodKeySignature

DidRemoval

This struct is passed as an argument while removing the DID did is the DID which is being removed.

DidSignature

GenesisConfig

Can be used to configure the genesis state of this pallet.

OffChainDidDetails

Stores details of an off-chain DID. Off-chain DID has no need of nonce as the signature is made on the whole transaction by the caller account and Substrate takes care of replay protection. Thus it stores the data about off-chain DID Doc (hash, URI or any other reference) and the account that owns it.

OnChainDidDetails

Stores details of an on-chain DID.

Pallet

The pallet implementing the on-chain logic.

RemoveControllers

RemoveKeys

RemoveServiceEndpoint

ServiceEndpoint

DID’s service endpoint.

ServiceEndpointId

DID’s service endpoint id.

ServiceEndpointOrigin

DID’s service endpoint origin.

ServiceEndpointWithId

ServiceEndpoint with its identifier.

UncheckedDidKey

DidKey without validity constraint requirement.

VerRelType

Different verification relation types specified in the DID spec here https://www.w3.org/TR/did-core/#verification-relationships.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

DidDetailsOrDidMethodKeyDetails

DidKeyOrDidMethodKey

Either DidKey or DidMethodKey.

DidMethodKey

The public_key in did:key:<public_key>.

DidOrDidMethodKey

Either did:dock:* or did:key:*.

DidOrDidMethodKeySignature

DID’s signature along with the used DIDs key reference or DidMethodKey’s signature.

Error

Error for the DID module.

Event

The event emitted by this pallet.

OffChainDidDocRef

To describe the off chain DID Doc’s reference. This is just to inform the client, this module does not check if the bytes are indeed valid as per the enum variant.

StoredDidDetails

Contains underlying DID describing its storage type.

Traits

Config

The module’s configuration trait.

HandleDidRemoval

Type Definitions

DidControllers

Stores controlled - controller pairs of a DID as (DID, DID) -> zero-sized record. If a record exists, then the controller is bound.

DidKeys

Stores keys of a DID as (DID, IncId) -> DidKey. Does not check if the same key is being added multiple times to the same DID.

DidMethodKeys

Stores nonce for did:key DIDs.

DidServiceEndpoints

Stores service endpoints of a DID as (DID, endpoint id) -> ServiceEndpoint.

Dids

Stores details of off-chain and on-chain DIDs

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

RawDid

Raw DID representation.

StoredOnChainDidDetails

Each on-chain DID is associated with a nonce that is incremented each time the DID does a write (through an extrinsic). The nonce starts from the block number when the DID was created to avoid replay attacks where an action of a DID that is removed and recreated by the same owner and same key is replayed by someone else.

Version