Trait dock_core_rpc::CoreModsApiServer

pub trait CoreModsApiServer<BlockHash, T>: Sized + Send + Sync + 'staticwhere
    T: ConfigWrapper,{

    // Required methods
    fn did_details<'life0, 'async_trait>(
        &'life0 self,
        did: Did,
        params: Option<AggregatedDidDetailsRequestParams>,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<AggregatedDidDetailsResponse<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn did_list_details<'life0, 'async_trait>(
        &'life0 self,
        dids: Vec<Did>,
        params: Option<AggregatedDidDetailsRequestParams>,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Option<AggregatedDidDetailsResponse<T::T>>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn bbs_public_key_with_params<'life0, 'async_trait>(
        &'life0 self,
        id: SignaturePublicKeyStorageKey,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<BBSPublicKeyWithParams<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn bbs_params_by_did<'life0, 'async_trait>(
        &'life0 self,
        owner: SignatureParamsOwner,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, BBSParameters<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn bbs_public_keys_by_did<'life0, 'async_trait>(
        &'life0 self,
        did: Did,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, BBSPublicKeyWithParams<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn bbs_plus_public_key_with_params<'life0, 'async_trait>(
        &'life0 self,
        id: SignaturePublicKeyStorageKey,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<BBSPlusPublicKeyWithParams<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn bbs_plus_params_by_did<'life0, 'async_trait>(
        &'life0 self,
        owner: SignatureParamsOwner,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, BBSPlusParameters<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn bbs_plus_public_keys_by_did<'life0, 'async_trait>(
        &'life0 self,
        did: Did,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, BBSPlusPublicKeyWithParams<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn ps_public_key_with_params<'life0, 'async_trait>(
        &'life0 self,
        id: SignaturePublicKeyStorageKey,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<PSPublicKeyWithParams<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn ps_params_by_did<'life0, 'async_trait>(
        &'life0 self,
        owner: SignatureParamsOwner,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, PSParameters<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn ps_public_keys_by_did<'life0, 'async_trait>(
        &'life0 self,
        did: Did,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, PSPublicKeyWithParams<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn accumulator_public_key_with_params<'life0, 'async_trait>(
        &'life0 self,
        id: AccumPublicKeyStorageKey,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<AccumPublicKeyWithParams<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn accumulator_with_public_key_and_params<'life0, 'async_trait>(
        &'life0 self,
        id: AccumulatorId,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<(Vec<u8>, Option<AccumPublicKeyWithParams<T::T>>)>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn schema_metadata<'life0, 'async_trait>(
        &'life0 self,
        id: TrustRegistrySchemaId,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistryId, AggregatedTrustRegistrySchemaMetadata<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn schema_issuers<'life0, 'async_trait>(
        &'life0 self,
        id: TrustRegistrySchemaId,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistryId, AggregatedTrustRegistrySchemaIssuers<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn schema_verifiers<'life0, 'async_trait>(
        &'life0 self,
        id: TrustRegistrySchemaId,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistryId, TrustRegistrySchemaVerifiers<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn schema_metadata_in_registry<'life0, 'async_trait>(
        &'life0 self,
        id: TrustRegistrySchemaId,
        registry_id: TrustRegistryId,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<AggregatedTrustRegistrySchemaMetadata<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn schema_issuers_in_registry<'life0, 'async_trait>(
        &'life0 self,
        id: TrustRegistrySchemaId,
        registry_id: TrustRegistryId,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<AggregatedTrustRegistrySchemaIssuers<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn schema_verifiers_in_registry<'life0, 'async_trait>(
        &'life0 self,
        id: TrustRegistrySchemaId,
        registry_id: TrustRegistryId,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<TrustRegistrySchemaVerifiers<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn all_registry_schema_metadata<'life0, 'async_trait>(
        &'life0 self,
        registry_id: TrustRegistryId,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistrySchemaId, AggregatedTrustRegistrySchemaMetadata<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn all_registry_schema_issuers<'life0, 'async_trait>(
        &'life0 self,
        registry_id: TrustRegistryId,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistrySchemaId, AggregatedTrustRegistrySchemaIssuers<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn all_registry_schema_verifiers<'life0, 'async_trait>(
        &'life0 self,
        registry_id: TrustRegistryId,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistrySchemaId, TrustRegistrySchemaVerifiers<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn registries_info_by<'life0, 'async_trait>(
        &'life0 self,
        by: QueryTrustRegistriesBy,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistryId, TrustRegistryInfo<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn registry_schemas_metadata_by<'life0, 'async_trait>(
        &'life0 self,
        by: QueryTrustRegistryBy,
        reg_id: TrustRegistryId,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistrySchemaId, AggregatedTrustRegistrySchemaMetadata<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn registries_ids_by<'life0, 'async_trait>(
        &'life0 self,
        by: QueryTrustRegistriesBy,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeSet<TrustRegistryId>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn registry_schemas_ids_by<'life0, 'async_trait>(
        &'life0 self,
        by: QueryTrustRegistryBy,
        reg_id: TrustRegistryId,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeSet<TrustRegistrySchemaId>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn bbdt16_public_key_with_params<'life0, 'async_trait>(
        &'life0 self,
        id: SignaturePublicKeyStorageKey,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<BBDT16PublicKeyWithParams<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn bbdt16_params_by_did<'life0, 'async_trait>(
        &'life0 self,
        owner: SignatureParamsOwner,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, BBDT16Parameters<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn bbdt16_public_keys_by_did<'life0, 'async_trait>(
        &'life0 self,
        did: Did,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, BBDT16PublicKeyWithParams<T::T>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self>
       where BlockHash: Send + Sync + 'static + DeserializeOwned,
             T: Send + Sync + 'static + Serialize + ConfigWrapper { ... }

}

Server trait implementation for the CoreModsApi RPC API.

Required Methods

fn did_details<'life0, 'async_trait>( &'life0 self, did: Did, params: Option<AggregatedDidDetailsRequestParams>, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<Option<AggregatedDidDetailsResponse<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn did_list_details<'life0, 'async_trait>( &'life0 self, dids: Vec<Did>, params: Option<AggregatedDidDetailsRequestParams>, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Option<AggregatedDidDetailsResponse<T::T>>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn bbs_public_key_with_params<'life0, 'async_trait>( &'life0 self, id: SignaturePublicKeyStorageKey, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<Option<BBSPublicKeyWithParams<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn bbs_params_by_did<'life0, 'async_trait>( &'life0 self, owner: SignatureParamsOwner, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, BBSParameters<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn bbs_public_keys_by_did<'life0, 'async_trait>( &'life0 self, did: Did, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, BBSPublicKeyWithParams<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn bbs_plus_public_key_with_params<'life0, 'async_trait>( &'life0 self, id: SignaturePublicKeyStorageKey, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<Option<BBSPlusPublicKeyWithParams<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn bbs_plus_params_by_did<'life0, 'async_trait>( &'life0 self, owner: SignatureParamsOwner, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, BBSPlusParameters<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn bbs_plus_public_keys_by_did<'life0, 'async_trait>( &'life0 self, did: Did, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, BBSPlusPublicKeyWithParams<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn ps_public_key_with_params<'life0, 'async_trait>( &'life0 self, id: SignaturePublicKeyStorageKey, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<Option<PSPublicKeyWithParams<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn ps_params_by_did<'life0, 'async_trait>( &'life0 self, owner: SignatureParamsOwner, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, PSParameters<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn ps_public_keys_by_did<'life0, 'async_trait>( &'life0 self, did: Did, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, PSPublicKeyWithParams<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn accumulator_public_key_with_params<'life0, 'async_trait>( &'life0 self, id: AccumPublicKeyStorageKey, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<Option<AccumPublicKeyWithParams<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn accumulator_with_public_key_and_params<'life0, 'async_trait>( &'life0 self, id: AccumulatorId, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<Option<(Vec<u8>, Option<AccumPublicKeyWithParams<T::T>>)>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn schema_metadata<'life0, 'async_trait>( &'life0 self, id: TrustRegistrySchemaId, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistryId, AggregatedTrustRegistrySchemaMetadata<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn schema_issuers<'life0, 'async_trait>( &'life0 self, id: TrustRegistrySchemaId, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistryId, AggregatedTrustRegistrySchemaIssuers<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn schema_verifiers<'life0, 'async_trait>( &'life0 self, id: TrustRegistrySchemaId, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistryId, TrustRegistrySchemaVerifiers<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn schema_metadata_in_registry<'life0, 'async_trait>( &'life0 self, id: TrustRegistrySchemaId, registry_id: TrustRegistryId, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<Option<AggregatedTrustRegistrySchemaMetadata<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn schema_issuers_in_registry<'life0, 'async_trait>( &'life0 self, id: TrustRegistrySchemaId, registry_id: TrustRegistryId, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<Option<AggregatedTrustRegistrySchemaIssuers<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn schema_verifiers_in_registry<'life0, 'async_trait>( &'life0 self, id: TrustRegistrySchemaId, registry_id: TrustRegistryId, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<Option<TrustRegistrySchemaVerifiers<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn all_registry_schema_metadata<'life0, 'async_trait>( &'life0 self, registry_id: TrustRegistryId, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistrySchemaId, AggregatedTrustRegistrySchemaMetadata<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn all_registry_schema_issuers<'life0, 'async_trait>( &'life0 self, registry_id: TrustRegistryId, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistrySchemaId, AggregatedTrustRegistrySchemaIssuers<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn all_registry_schema_verifiers<'life0, 'async_trait>( &'life0 self, registry_id: TrustRegistryId, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistrySchemaId, TrustRegistrySchemaVerifiers<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn registries_info_by<'life0, 'async_trait>( &'life0 self, by: QueryTrustRegistriesBy, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistryId, TrustRegistryInfo<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn registry_schemas_metadata_by<'life0, 'async_trait>( &'life0 self, by: QueryTrustRegistryBy, reg_id: TrustRegistryId, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<TrustRegistrySchemaId, AggregatedTrustRegistrySchemaMetadata<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn registries_ids_by<'life0, 'async_trait>( &'life0 self, by: QueryTrustRegistriesBy, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeSet<TrustRegistryId>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn registry_schemas_ids_by<'life0, 'async_trait>( &'life0 self, by: QueryTrustRegistryBy, reg_id: TrustRegistryId, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeSet<TrustRegistrySchemaId>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn bbdt16_public_key_with_params<'life0, 'async_trait>( &'life0 self, id: SignaturePublicKeyStorageKey, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<Option<BBDT16PublicKeyWithParams<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn bbdt16_params_by_did<'life0, 'async_trait>( &'life0 self, owner: SignatureParamsOwner, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, BBDT16Parameters<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn bbdt16_public_keys_by_did<'life0, 'async_trait>( &'life0 self, did: Did, at: Option<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<IncId, BBDT16PublicKeyWithParams<T::T>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Provided Methods

fn into_rpc(self) -> RpcModule<Self>where BlockHash: Send + Sync + 'static + DeserializeOwned, T: Send + Sync + 'static + Serialize + ConfigWrapper,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.